requirements.txt
async-generator==1.10
attrs==22.2.0
black==23.1.0
Brotli==1.0.9
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
ConfigArgParse==1.5.3
coverage==7.2.0
Flask==2.2.3
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
gevent==22.10.2
geventhttpclient==2.0.8
greenlet==2.0.2
h11==0.14.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
locust==2.14.2
MarkupSafe==2.1.2
msgpack==1.0.4
mypy-extensions==1.0.0
outcome==1.2.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
pluggy==1.0.0
psutil==5.9.4
PySocks==1.7.1
pytest==7.2.1
pytest-cov==4.0.0
pytest-flask==1.2.0
python-dotenv==0.21.1
pyzmq==25.0.0
requests==2.28.2
roundrobin==0.0.4
selenium==4.8.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
wsproto==1.2.0
zope.event==4.6
zope.interface==5.5.2
